+++ CHANGES.rst
@@ -2,6 +2,12 @@ Weitersager Changelog
 =====================
 
 
+Version 0.4
+-----------
+
+Unreleased
+
+
 Version 0.3
 -----------